поправка помарок, исправление тестов и выделение кейсов тестов в handlers, объединение мелких файлов
