+++ uniq/handlers/handlers_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 )
 
-func TestHandleLines(t *testing.T) {
+func TestWithoutOptions(t *testing.T) {
 	input := []string{
 		"I love music.",
 		"I love music.",
@@ -30,7 +30,21 @@ func TestHandleLines(t *testing.T) {
 	if !reflect.DeepEqual(HandleLines(input, opts), expResultWithoutOptions) {
 		t.Fatalf("wrong result")
 	}
-	opts.EShowStrMeetCount = true
+}
+
+func TestStrMeetCountFlag(t *testing.T) {
+	input := []string{
+		"I love music.",
+		"I love music.",
+		"I love music.",
+		"",
+		"I love music of Kartik.",
+		"I love music of Kartik.",
+		"Thanks.",
+		"I love music of Kartik.",
+		"I love music of Kartik.",
+	}
+	opts := options.Options{EShowStrMeetCount: true}
 
 	expResultWithCFlag := []string{
 		"3 I love music.",
@@ -43,9 +57,21 @@ func TestHandleLines(t *testing.T) {
 	if !reflect.DeepEqual(HandleLines(input, opts), expResultWithCFlag) {
 		t.Fatalf("wrong result")
 	}
+}
 
-	opts.EShowStrMeetCount = false
-	opts.EShowNotUniqueStr = true
+func TestShowNotUniqueStr(t *testing.T) {
+	input := []string{
+		"I love music.",
+		"I love music.",
+		"I love music.",
+		"",
+		"I love music of Kartik.",
+		"I love music of Kartik.",
+		"Thanks.",
+		"I love music of Kartik.",
+		"I love music of Kartik.",
+	}
+	opts := options.Options{EShowNotUniqueStr: true}
 
 	expResultWithDFlag := []string{
 		"I love music.",
@@ -56,8 +82,22 @@ func TestHandleLines(t *testing.T) {
 	if !reflect.DeepEqual(HandleLines(input, opts), expResultWithDFlag) {
 		t.Fatalf("wrong result")
 	}
-	opts.EShowNotUniqueStr = false
-	opts.EShowUniqueStr = true
+}
+
+func TestShowUniqueStr(t *testing.T) {
+	input := []string{
+		"I love music.",
+		"I love music.",
+		"I love music.",
+		"",
+		"I love music of Kartik.",
+		"I love music of Kartik.",
+		"Thanks.",
+		"I love music of Kartik.",
+		"I love music of Kartik.",
+	}
+	opts := options.Options{EShowUniqueStr: true}
+
 	expResultWithUFlag := []string{
 		"",
 		"Thanks.",
@@ -65,6 +105,9 @@ func TestHandleLines(t *testing.T) {
 	if !reflect.DeepEqual(HandleLines(input, opts), expResultWithUFlag) {
 		t.Fatalf("wrong result")
 	}
+}
+
+func TestIgnoreRegisterFlag(t *testing.T) {
 	inputIFlag := []string{
 		"I LOVE MUSIC.",
 		"I love music.",
@@ -76,8 +119,8 @@ func TestHandleLines(t *testing.T) {
 		"I love music of kartik.",
 		"I love MuSIC of Kartik.",
 	}
-	opts.EShowUniqueStr = false
-	opts.IgnoreRegister = true
+	opts := options.Options{IgnoreRegister: true}
+
 	expResultWithIFlag := []string{
 		"I LOVE MUSIC.",
 		"",
@@ -88,7 +131,9 @@ func TestHandleLines(t *testing.T) {
 	if !reflect.DeepEqual(HandleLines(inputIFlag, opts), expResultWithIFlag) {
 		t.Fatalf("wrong result")
 	}
+}
 
+func TestIgnoreNumFields(t *testing.T) {
 	inputFFlag := []string{
 		"We love music.",
 		"I love music.",
@@ -104,12 +149,13 @@ func TestHandleLines(t *testing.T) {
 		"I love music of Kartik.",
 		"Thanks.",
 	}
-	opts.IgnoreRegister = false
-	opts.SkippedStringsCount = 1
+	opts := options.Options{SkippedStringsCount: 1}
 	if !reflect.DeepEqual(HandleLines(inputFFlag, opts), expResultWithFFlag) {
 		t.Fatalf("wrong result")
 	}
+}
 
+func TestIgnoreNumChars(t *testing.T) {
 	inputSFlag := []string{
 		"I love music.",
 		"A love music.",
@@ -126,10 +172,9 @@ func TestHandleLines(t *testing.T) {
 		"We love music of Kartik.",
 		"Thanks.",
 	}
-	opts.SkippedStringsCount = 0
-	opts.SkippedCharsCount = 1
-
+	opts := options.Options{SkippedCharsCount: 1}
 	if !reflect.DeepEqual(HandleLines(inputSFlag, opts), expResultWithSFlag) {
 		t.Fatalf("wrong result")
 	}
+
 }
