--- uniq/handlers/show_str_met_count.go
@@ -1,20 +1,0 @@
-package handlers
-
-import (
-	"fmt"
-
-	"GolangCourse/uniq/options"
-	"GolangCourse/uniq/utils"
-)
-
-func formatMeetCountResult(linesWithMetInfo []utils.StringWithMetCount) (result []string) {
-	for _, strWithMetInfo := range linesWithMetInfo {
-		result = append(result, fmt.Sprintf("%d %s", strWithMetInfo.MetCount, strWithMetInfo.Str))
-	}
-	return result
-}
-
-func showStrMeetCount(lines []string, opts options.Options) (resultLines []string) {
-	linesWithMetInfo := utils.GetStringsWithMetCount(lines, opts.IgnoreRegister, opts.SkippedStringsCount, opts.SkippedCharsCount)
-	return formatMeetCountResult(linesWithMetInfo)
-}
