--- uniq/utils/skip_chars.go
@@ -1,8 +1,0 @@
-package utils
-
-func SkipChars(line string, skipCharsCount int) (resultLine string) {
-	if len(line) < skipCharsCount {
-		return resultLine
-	}
-	return line[skipCharsCount:]
-}
