--- uniq/utils/skip_words.go
@@ -1,20 +1,0 @@
-package utils
-
-import (
-	"fmt"
-	"regexp"
-	"strings"
-)
-
-func getWordsCount(line string) int {
-	return strings.Count(strings.Trim(line, " "), " ") + 1
-}
-
-func SkipWords(line string, skipWordsCount int) (resultLine string) {
-	wordsCount := getWordsCount(line)
-	if wordsCount <= skipWordsCount {
-		return resultLine
-	}
-	re := regexp.MustCompile(fmt.Sprintf("^(.*? ){%d}", skipWordsCount))
-	return re.ReplaceAllString(strings.TrimLeft(line, " "), "")
-}
