--- uniq/utils/skip_words_and_chars.go
@@ -1,5 +1,0 @@
-package utils
-
-func SkipWordsAndChars(line string, skipWordsCount, skipCharsCount int) string {
-	return SkipChars(SkipWords(line, skipWordsCount), skipCharsCount)
-}
