--- uniq/utils/string_compare_func.go
@@ -1,12 +1,0 @@
-package utils
-
-import "strings"
-
-func GetStringCompareFunc(ignoreRegister bool) func(lhs string, rhs string) bool {
-	if ignoreRegister {
-		return strings.EqualFold
-	}
-	return func(lhs string, rhs string) bool {
-		return lhs == rhs
-	}
-}
